Binary Search - M - 13

+++ DSA/BinarySearch/Medium/13_MedianOf2SortedArrays.java
@@ -109,9 +109,9 @@ class Main {
             }
 
             if(totalLength%2 == 0){
-                return (double) ((double)(prev+curr)/2);
+                return (double) (prev+curr)/2;
             }
-            
+
             return (double) curr;
         }
